requirements.txt
streamlit>=1.40.0
pymongo[srv]>=4.7.2
certifi>=2024.0.0
bcrypt>=4.1.3
python-dotenv>=1.0.1
Pillow>=10.2.0
pandas>=2.2.1
plotly>=5.20.0
fpdf2>=2.7.9
qrcode>=7.4.2
python-dateutil>=2.9.0
